Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/parsing/column-width-computed-expected.txt


PASS Property column-width value 'auto' computes to 'auto' 
PASS Property column-width value 'calc(10px + 0.5em)' computes to '30px' 
FAIL Property column-width value 'calc(10px - 0.5em)' computes to '0px' assert_equals: computed value should round-trip expected "0px" but got "auto"
